Refactor handling of the ToD manager
Instead of the individual classes such as game_board or map_context owning a manager, we instead use a singleton-like
design where the manager is reset or cleared by individual classes, such as the two above, during their lifetime.

This also removes resources::tod_manager in favor of the singleton.