Lua: Refactor the AI module to be dir()-friendly
This places all read-only functions straight into the AI table, which means it sticks when stdlib.lua nulls out the old move map API.

The AI table then implements the __dir metamethod to add the mutating functions in if read_only is false.

This also adds a __dir metamethod for the aspects table which gets the list of aspects directly from the engine.